Add files via upload
This is my Python code for the sales analysis project for Value Inc. company. 

Value Inc is a retail store that sells household items all over the world in bulk.
The aim of the analysis has been done is to create a dashboard on the monthly cost, profit, and top-selling products and generally give a sales report on different aspects of the business. 

Data file is used for this project is csv. file and consisting of 1,047,588 rows and 19 columns.